AddArticles

+++ src/main/java/com/shit/Delete.java
@@ -13,6 +13,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.sql.SQLException;
 
 @WebServlet("/delete")
@@ -20,19 +21,24 @@ public class Delete extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        resp.setContentType("text/html;charset=UTF-8");
+        req.setCharacterEncoding("UTF-8");
         HttpSession session = req.getSession();
+        User user = (User) session.getAttribute("user");
 
-        User user;
-        String username = (String) session.getAttribute("username");
-        UserDao userDao = new UserDao(DataConnection.getConnection());
-        try  {
-            user = userDao.getUserByUsername(username);
-            userDao.delete(user);
-        } catch (SQLException e) {
-            e.printStackTrace();
+        if (user.getRoleId() == 1 && (!user.getUsername().equals(req.getParameter("username")))){
+            UserDao userDao = new UserDao(DataConnection.getConnection());
+            try {
+                user = userDao.getUserByUsername(req.getParameter("username"));
+                userDao.delete(user);
+            } catch (SQLException e) {
+                System.out.println("Не удалось найти данного пользователя!");
+            } finally {
+                req.getRequestDispatcher("/edit").forward(req, resp);
+            }
+        } else {
+            PrintWriter pw = resp.getWriter();
+            pw.print("Данная операция невозможна!");
         }
-        session.getAttribute("title");
-        req.getRequestDispatcher("index.jsp").forward(req, resp);
-
     }
 }
